Fixed some bugs @3 , Passed Wrong Parameter from repo to dao

+++ src/main/kotlin/instaU/ayush/com/dao/qna/QnaDaoImpl.kt
@@ -60,7 +60,7 @@ class QnaDaoImpl : QnaDao {
         return dbQuery {
             val insertStatement = AnswersTable.insert {
                 it[AnswersTable.answerId] = IdGenerator.generateId()
-                it[AnswersTable.questionId] = questionId.toLong()
+                it[AnswersTable.questionId] = questionId
                 it[AnswersTable.userId] = authorId
                 it[AnswersTable.answer] = content
             }
